[Refactor] Move BackendSelector to files
