+++ fe/fe-core/src/main/java/com/starrocks/qe/CoordinatorPreprocessor.java
@@ -76,12 +76,10 @@ import com.starrocks.thrift.TQueryGlobals;
 import com.starrocks.thrift.TQueryOptions;
 import com.starrocks.thrift.TQueryType;
 import com.starrocks.thrift.TRuntimeFilterParams;
-import com.starrocks.thrift.TScanRangeLocation;
 import com.starrocks.thrift.TScanRangeLocations;
 import com.starrocks.thrift.TScanRangeParams;
 import com.starrocks.thrift.TUniqueId;
 import com.starrocks.thrift.TWorkGroup;
-import org.apache.commons.collections4.MapUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -1705,266 +1703,6 @@ public class CoordinatorPreprocessor {
             appendTo(sb);
             return sb.toString();
         }
-    }
-
-    public static class NormalBackendSelector implements BackendSelector {
-        private final ScanNode scanNode;
-        private final List<TScanRangeLocations> locations;
-        private final FragmentScanRangeAssignment assignment;
-
-        private final WorkerProvider workerProvider;
-        private final boolean isLoad;
-
-        public NormalBackendSelector(ScanNode scanNode, List<TScanRangeLocations> locations,
-                                     FragmentScanRangeAssignment assignment, WorkerProvider workerProvider,
-                                     boolean isLoad) {
-            this.scanNode = scanNode;
-            this.locations = locations;
-            this.assignment = assignment;
-            this.workerProvider = workerProvider;
-            this.isLoad = isLoad;
-        }
-
-        private boolean isEnableScheduleByRowCnt(TScanRangeLocations scanRangeLocations) {
-            // only enable for load now, The insert into select performance problem caused by data skew is the most serious
-            return Config.enable_schedule_insert_query_by_row_count && isLoad
-                    && scanRangeLocations.getScan_range().isSetInternal_scan_range()
-                    && scanRangeLocations.getScan_range().getInternal_scan_range().isSetRow_count()
-                    && scanRangeLocations.getScan_range().getInternal_scan_range().getRow_count() > 0;
-        }
-
-        @Override
-        public void computeScanRangeAssignment() throws UserException {
-            HashMap<TNetworkAddress, Long> assignedRowCountPerHost = Maps.newHashMap();
-            // sort the scan ranges by row count
-            // only sort the scan range when it is load job
-            // but when there are too many scan ranges, we will not sort them since performance issue
-            if (locations.size() < 10240 && !locations.isEmpty() && isEnableScheduleByRowCnt(locations.get(0))) {
-                locations.sort((lhs, rhs) -> Long.compare(
-                        rhs.getScan_range().getInternal_scan_range().getRow_count(),
-                        lhs.getScan_range().getInternal_scan_range().getRow_count()));
-            }
-
-            for (TScanRangeLocations scanRangeLocations : locations) {
-                // assign this scan range to the host w/ the fewest assigned row count
-                Long minAssignedBytes = Long.MAX_VALUE;
-                TScanRangeLocation minLocation = null;
-                for (final TScanRangeLocation location : scanRangeLocations.getLocations()) {
-                    if (!workerProvider.isDataNodeAvailable(location.getBackend_id())) {
-                        continue;
-                    }
-
-                    Long assignedBytes = assignedRowCountPerHost.getOrDefault(location.server, 0L);
-                    if (assignedBytes < minAssignedBytes) {
-                        minAssignedBytes = assignedBytes;
-                        minLocation = location;
-                    }
-                }
-
-                if (minLocation == null) {
-                    workerProvider.reportDataNodeNotFoundException();
-                }
-                Preconditions.checkNotNull(minLocation);
-
-                // only enable for load now, The insert into select performance problem caused by data skew is the most serious
-                if (isEnableScheduleByRowCnt(scanRangeLocations)) {
-                    assignedRowCountPerHost.put(minLocation.server, minAssignedBytes
-                            + scanRangeLocations.getScan_range().getInternal_scan_range().getRow_count());
-                } else {
-                    // use tablet num as assigned row count
-                    assignedRowCountPerHost.put(minLocation.server, minAssignedBytes + 1);
-                }
-                workerProvider.selectWorker(minLocation.backend_id);
-                // add scan range
-                TScanRangeParams scanRangeParams = new TScanRangeParams(scanRangeLocations.scan_range);
-                assignment.put(minLocation.backend_id, scanNode.getId().asInt(), scanRangeParams);
-            }
-
-            if (LOG.isDebugEnabled()) {
-                LOG.debug("assignedRowCountPerHost: {}", assignedRowCountPerHost);
-            }
-        }
-    }
-
-    public static class ReplicatedBackendSelector implements BackendSelector {
-        private final ScanNode scanNode;
-        private final List<TScanRangeLocations> locations;
-        private final FragmentScanRangeAssignment assignment;
-        private final ColocatedBackendSelector.Assignment colocatedAssignment;
-
-        public ReplicatedBackendSelector(ScanNode scanNode, List<TScanRangeLocations> locations,
-                                         FragmentScanRangeAssignment assignment,
-                                         ColocatedBackendSelector.Assignment colocatedAssignment) {
-            this.scanNode = scanNode;
-            this.locations = locations;
-            this.assignment = assignment;
-            this.colocatedAssignment = colocatedAssignment;
-        }
-
-        @Override
-        public void computeScanRangeAssignment() {
-            for (TScanRangeLocations scanRangeLocations : locations) {
-                for (Map<Integer, List<TScanRangeParams>> scanRanges : assignment.values()) {
-                    List<TScanRangeParams> scanRangeParamsList =
-                            scanRanges.computeIfAbsent(scanNode.getId().asInt(), k -> Lists.newArrayList());
-                    // add scan range
-                    TScanRangeParams scanRangeParams = new TScanRangeParams(scanRangeLocations.scan_range);
-                    scanRangeParamsList.add(scanRangeParams);
-                }
-            }
-            // If this fragment has bucket/colocate join, there need to fill fragmentIdBucketSeqToScanRangeMap here.
-            // For example:
-            //                       join(replicated)
-            //                    /                    \
-            //            join(bucket/colocate)       scan(C)
-            //              /           \
-            //            scan(A)         scan(B)
-            // There are replicate join and bucket/colocate join in same fragment. for each bucket A,B used, we need to
-            // add table C all tablet because of the character of the replicate join.
-            if (colocatedAssignment != null && !MapUtils.isEmpty(colocatedAssignment.getSeqToScanRange())) {
-                ColocatedBackendSelector.BucketSeqToScanRange bucketSeqToScanRange =
-                        colocatedAssignment.getSeqToScanRange();
-                for (Map.Entry<Integer, Map<Integer, List<TScanRangeParams>>> entry : bucketSeqToScanRange.entrySet()) {
-                    for (TScanRangeLocations scanRangeLocations : locations) {
-                        List<TScanRangeParams> scanRangeParamsList =
-                                entry.getValue().computeIfAbsent(scanNode.getId().asInt(), k -> Lists.newArrayList());
-                        // add scan range
-                        TScanRangeParams scanRangeParams = new TScanRangeParams(scanRangeLocations.scan_range);
-                        scanRangeParamsList.add(scanRangeParams);
-                    }
-                }
-            }
-        }
-    }
-
-    public static class ColocatedBackendSelector implements BackendSelector {
-
-        private final OlapScanNode scanNode;
-        private final FragmentScanRangeAssignment assignment;
-        private final ColocatedBackendSelector.Assignment colocatedAssignment;
-        private final boolean isRightOrFullBucketShuffleFragment;
-        private final WorkerProvider workerProvider;
-
-        public ColocatedBackendSelector(OlapScanNode scanNode, FragmentScanRangeAssignment assignment,
-                                        ColocatedBackendSelector.Assignment colocatedAssignment,
-                                        boolean isRightOrFullBucketShuffleFragment, WorkerProvider workerProvider) {
-            this.scanNode = scanNode;
-            this.assignment = assignment;
-            this.colocatedAssignment = colocatedAssignment;
-            this.isRightOrFullBucketShuffleFragment = isRightOrFullBucketShuffleFragment;
-            this.workerProvider = workerProvider;
-        }
-
-        @Override
-        public void computeScanRangeAssignment() throws UserException {
-            Map<Integer, Long> bucketSeqToWorkerId = colocatedAssignment.seqToWorkerId;
-            ColocatedBackendSelector.BucketSeqToScanRange bucketSeqToScanRange =
-                    colocatedAssignment.seqToScanRange;
-
-            for (Integer bucketSeq : scanNode.bucketSeq2locations.keySet()) {
-                List<TScanRangeLocations> locations = scanNode.bucketSeq2locations.get(bucketSeq);
-                if (!bucketSeqToWorkerId.containsKey(bucketSeq)) {
-                    computeExecAddressForBucketSeq(locations.get(0), bucketSeq);
-                }
-
-                List<TScanRangeParams> scanRangeParamsList =
-                        bucketSeqToScanRange
-                                .computeIfAbsent(bucketSeq, k -> Maps.newHashMap())
-                                .computeIfAbsent(scanNode.getId().asInt(), k -> Lists.newArrayList());
-                locations.stream()
-                        .map(location -> new TScanRangeParams(location.scan_range))
-                        .forEach(scanRangeParamsList::add);
-            }
-            // Because of the right table will not send data to the bucket which has been pruned, the right join or full join will get wrong result.
-            // So if this bucket shuffle is right join or full join, we need to add empty bucket scan range which is pruned by predicate.
-            if (isRightOrFullBucketShuffleFragment) {
-                int bucketNum = colocatedAssignment.bucketNum;
-
-                for (int bucketSeq = 0; bucketSeq < bucketNum; ++bucketSeq) {
-                    if (!bucketSeqToWorkerId.containsKey(bucketSeq)) {
-                        long workerId = workerProvider.selectNextWorker();
-                        bucketSeqToWorkerId.put(bucketSeq, workerId);
-                    }
-                    if (!bucketSeqToScanRange.containsKey(bucketSeq)) {
-                        bucketSeqToScanRange.put(bucketSeq, Maps.newHashMap());
-                        bucketSeqToScanRange.get(bucketSeq).put(scanNode.getId().asInt(), Lists.newArrayList());
-                    }
-                }
-            }
-
-            // use bucketSeqToScanRange to fill FragmentScanRangeAssignment
-            bucketSeqToScanRange.forEach((seq, nodeId2ScanRanges) -> {
-                // fill FragmentScanRangeAssignment only when there are scan id in the bucket
-                int scanNodeId = scanNode.getId().asInt();
-                if (nodeId2ScanRanges.containsKey(scanNodeId)) {
-                    assignment.putAll(bucketSeqToWorkerId.get(seq), scanNodeId, nodeId2ScanRanges.get(scanNodeId));
-                }
-            });
-        }
-
-        // Make sure each host have average bucket to scan
-        private void computeExecAddressForBucketSeq(TScanRangeLocations seqLocation, Integer bucketSeq)
-                throws UserException {
-            Map<Long, Integer> buckendIdToBucketCountMap = colocatedAssignment.backendIdToBucketCount;
-            int minBucketNum = Integer.MAX_VALUE;
-            long minBackendId = Long.MAX_VALUE;
-            for (TScanRangeLocation location : seqLocation.locations) {
-                if (!workerProvider.isDataNodeAvailable(location.getBackend_id())) {
-                    continue;
-                }
-
-                Integer bucketNum = buckendIdToBucketCountMap.getOrDefault(location.backend_id, 0);
-                if (bucketNum < minBucketNum) {
-                    minBucketNum = bucketNum;
-                    minBackendId = location.backend_id;
-                }
-            }
-
-            if (minBackendId == Long.MAX_VALUE) {
-                workerProvider.reportDataNodeNotFoundException();
-            }
-
-            buckendIdToBucketCountMap.put(minBackendId, minBucketNum + 1);
-            workerProvider.selectWorker(minBackendId);
-            colocatedAssignment.seqToWorkerId.put(bucketSeq, minBackendId);
-        }
-
-        public static class BucketSeqToScanRange
-                extends HashMap<Integer, Map<Integer, List<TScanRangeParams>>> {
-        }
-
-        public static class Assignment {
-            private final Map<Integer, Long> seqToWorkerId = Maps.newHashMap();
-            // < bucket_seq -> < scannode_id -> scan_range_params >>
-            private final ColocatedBackendSelector.BucketSeqToScanRange seqToScanRange =
-                    new ColocatedBackendSelector.BucketSeqToScanRange();
-            private final Map<Long, Integer> backendIdToBucketCount = Maps.newHashMap();
-            private final int bucketNum;
-
-            public Assignment(OlapScanNode scanNode) {
-                int curBucketNum = scanNode.getOlapTable().getDefaultDistributionInfo().getBucketNum();
-                if (scanNode.getSelectedPartitionIds().size() <= 1) {
-                    for (Long pid : scanNode.getSelectedPartitionIds()) {
-                        curBucketNum = scanNode.getOlapTable().getPartition(pid).getDistributionInfo().getBucketNum();
-                    }
-                }
-
-                this.bucketNum = curBucketNum;
-            }
-
-            public Map<Integer, Long> getSeqToWorkerId() {
-                return seqToWorkerId;
-            }
-
-            public ColocatedBackendSelector.BucketSeqToScanRange getSeqToScanRange() {
-                return seqToScanRange;
-            }
-
-            public int getBucketNum() {
-                return bucketNum;
-            }
-        }
-
     }
 
     public TNetworkAddress getAddressByWorkerId(long workerId) {
